requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
mysql==0.0.3
mysqlclient==2.2.7
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
